test: enable InlineEditorProviders-integ-test
